Move wire protocol types to separate files
